Increase the test coverage. The test suite is broken because of views.newsroom.save_page. Not sure what is the expected behavior of this view

+++ newsroom_project/apps/stories/views/newsroom.py
@@ -63,28 +63,31 @@ def save_page(request,story_id):
     saves (creates or replaces) a story
     """
     story = get_object_or_404(Story,pk=story_id)
-    page_formset = PageFormSet(request.POST)
-    if page_formset.is_valid():
-        for form in page_formset.cleaned_data:
-            page = None
-            try:
-                page = Page.objects.get(story=story,pagenum=form['pagenum'])
-            except MultipleObjectsReturned:
-                Page.objects.filter(story=story,pagenum=form['pagenum']).delete()
-            except ObjectDoesNotExist:
-                pass
+    if request.POST:
+        page_formset = PageFormSet(request.POST)
+        if page_formset.is_valid():
 
-            if not page:
-                page = Page()
-                page.story = story
-                page.pagenum = form['pagenum']
+            for form in page_formset.forms:
+                import ipdb; ipdb.set_trace()
+                page = None
+                try:
+                    page = Page.objects.get(story=story,pagenum=form['pagenum'])
+                except MultipleObjectsReturned:
+                    Page.objects.filter(story=story,pagenum=form['pagenum']).delete()
+                except ObjectDoesNotExist:
+                    pass
 
-            page.content = form['content']
-            page.save()
-        return HttpResponse("1", mimetype="text/plain")
-    else:
-        print page_formset.errors
-        return HttpResponse("-1", mimetype="text/plain")
+                if not page:
+                    page = Page()
+                    page.story = story
+                    page.pagenum = form['pagenum']
+
+                page.content = form['content']
+                page.save()
+            return HttpResponse("1", mimetype="text/plain")
+        else:
+            print page_formset.errors
+            return HttpResponse("-1", mimetype="text/plain")
 
 @login_required
 def edit_story(request,story_id):
